pyproject.toml
[project]
name = "fastapi_test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "fastapi>=0.116.1",
    "httpie>=3.2.4",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]


[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["app/test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
